Remove "Unified Address" from YWallet features
YWallet does not fully implement ZIP 316, the Unified Address specification. It differs from the specification both in its treatment of transparent address generation, and in that it does not handle change being sent to the wallet's internal keys; restoring a seed from a ZIP 316-compliant wallet will not show the correct balance, because all change outputs will be missing. For example, using the same seed, my YWallet balance is almost zero whereas the correct balance (as seen in Zashi) is ~4 ZEC. I have not checked the behavior of other wallets; I believe this might be a problem for Zingo as well.